requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.4.1
    # via
    #   django
    #   djoser
certifi==2021.5.30
    # via requests
cffi==1.14.6
    # via cryptography
charset-normalizer==2.0.4
    # via requests
coreapi==2.3.3
    # via djoser
coreschema==0.0.4
    # via coreapi
cryptography==3.4.8
    # via social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==3.2.6
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-extra-fields
django-cors-headers==3.8.0
    # via -r requirements.in
django-environ==0.4.5
    # via -r requirements.in
django-filter==2.4.0
    # via -r requirements.in
django-split-settings==1.1.0
    # via -r requirements.in
django-templated-mail==1.1.1
    # via djoser
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-extra-fields
djangorestframework-simplejwt==4.8.0
    # via djoser
djoser==2.1.0
    # via -r requirements.in
drf-extra-fields==3.1.1
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.2
    # via requests
itypes==1.2.0
    # via coreapi
jinja2==3.0.1
    # via coreschema
markupsafe==2.0.1
    # via jinja2
oauthlib==3.1.1
    # via
    #   requests-oauthlib
    #   social-auth-core
pillow==8.3.1
    # via
    #   -r requirements.in
    #   reportlab
psycopg2-binary==2.9.1
    # via -r requirements.in
pycparser==2.20
    # via cffi
pyjwt==2.1.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
python3-openid==3.2.0
    # via social-auth-core
pytz==2021.1
    # via django
reportlab==3.6.1
    # via -r requirements.in
requests==2.26.0
    # via
    #   coreapi
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.0
    # via social-auth-core
six==1.16.0
    # via social-auth-app-django
social-auth-app-django==4.0.0
    # via djoser
social-auth-core==4.1.0
    # via social-auth-app-django
sqlparse==0.4.1
    # via django
uritemplate==3.0.1
    # via coreapi
urllib3==1.26.6
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
